proto: makes box coordinates as short

+++ src/client/video/writer.py
@@ -53,9 +53,9 @@ class AsyncDetectionWriter:
 
 @dataclasses.dataclass
 class DetectionResult:
-    x1: float
-    y1: float
-    x2: float
-    y2: float
+    x1: int
+    y1: int
+    x2: int
+    y2: int
     score: float
     clazz: int
